Move hot.ts into src/common
