+++ src/runtime/crdt/crdt-collection.ts
@@ -9,7 +9,7 @@
  */
 
 import {ChangeType, CRDTChange, CRDTError, CRDTModel, CRDTTypeRecord, VersionMap} from './crdt.js';
-import {Dictionary} from '../hot.js';
+import {Dictionary} from '../../common/hot.js';
 
 type RawCollection<T> = Set<T>;
