+++ src/runtime/crdt/crdt.ts
@@ -8,7 +8,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
- import {Dictionary} from '../hot.js';
+ import {Dictionary} from '../../common/hot.js';
 
 export type VersionMap = Dictionary<number>;
