+++ src/runtime/storage/storage-provider-base.ts
@@ -17,7 +17,7 @@ import {ModelValue, SerializedModelEntry} from './crdt-collection-model.js';
 import {KeyBase} from './key-base.js';
 import {Store, BigCollectionStore, CollectionStore, SingletonStore} from '../store.js';
 import {PropagatedException} from '../arc-exceptions.js';
-import {Dictionary, Consumer} from '../hot.js';
+import {Dictionary, Consumer} from '../../common/hot.js';
 import {ClaimIsTag} from '../particle-claim.js';
 
 enum EventKind {
